Fix example
