+++ example/index.html
@@ -3,29 +3,10 @@
 <head>
   <meta charset="UTF-8">
   <title>Example</title>
-  <style>
-    body {
-      font-family: Helvetica;
-    }
-    input {
-      display: block;
-    }
-    img {
-      width: 300px;
-      border: 1px solid;
-      margin: 5px;
-    }
-    video {
-      width: 500px;
-      display: block;
-      margin: 10px;
-      border: 3px solid;
-    }
-  </style>
   <!-- <script src="https://cdn.polyfill.io/v2/polyfill.js?features=es6,Array.prototype.includes,fetch"></script> -->
 </head>
 <body>
-  <input id="browse" type="file">
-  <script src="../index.js"></script>
+  <div id="app"></div>
+  <script src="../example-bundle.js"></script>
 </body>
 </html>
