+++ example/styled.ts
@@ -1,4 +1,24 @@
-import styled from 'styled-components';
+import styled, {injectGlobal} from 'styled-components';
+
+injectGlobal`
+  body {
+    font-family: Helvetica;
+  }
+  input {
+    display: block;
+  }
+  img {
+    width: 300px;
+    border: 1px solid;
+    margin: 5px;
+  }
+  video {
+    width: 500px;
+    display: block;
+    margin: 10px;
+    border: 3px solid;
+  }
+`;
 
 export const Preview = styled.img`
   width: 300px;
